requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
PyMySQL==1.1.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.0.1

# Configuration
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# External APIs
openai==1.12.0
httpx==0.26.0

# Development
python-dotenv==1.0.1
